fix: db errors when deleting a table that doesn't exist

+++ gradebookApp/models/Class_list_model.php
@@ -128,24 +128,28 @@ class Class_list_model extends \MY_Model
             ->where("class_id", $classObj->class_id)
             ->delete("students_enrolled");
 
-        /**
-         * Deletes assignments contained within class_table
-         */
-        $results = $this->db
-            ->select("assignment_id as id")
-            ->from($classObj->table_name)
-            ->get()->result_array();
-        $results = array_unique($results, SORT_REGULAR);
-        foreach ($results as $result) {
-            $this->db->or_where($result);
-        }
-        $this->db->delete("assignments");
+        if ($this->db->table_exists($classObj->table_name)) {
+            /**
+             * Deletes assignments contained within class_table
+             */
+            $results = $this->db
+                ->select("assignment_id as id")
+                ->from($classObj->table_name)
+                ->get()->result_array();
+            $results = array_unique($results, SORT_REGULAR);
+            foreach ($results as $result) {
+                $this->db->or_where($result);
+            }
+            if (count($results) > 0) {
+                $this->db->delete("assignments");
+            }
 
-        /**
-         * Drops table for class from db
-         */
-        $this->load->dbforge();
-        $this->dbforge->drop_table($classObj->table_name, true);
+            /**
+             * Drops table for class from db
+             */
+            $this->load->dbforge();
+            $this->dbforge->drop_table($classObj->table_name, true);
+        }
     }
 
     /**
